src/main/java/com/zipcodewilmington/Person.java
package com.zipcodewilmington;

/**
 * Created by leon on 1/24/18.
 */
public class Person {
    private final String firstName;
    private final String lastName;
    private Integer age;
    private String height;
    private Double weight;

    public Person(String firstName, String lastName) {
        this.firstName = firstName;
        this.lastName = lastName;
    }
    public Person(String firstName, String lastName, Integer newAge, String newHeight, Double newWeight){
        this.firstName = firstName;
        this.lastName = lastName;
        this.age = newAge;
        this.height = newHeight;
        this.weight = newWeight;
    }

    public String getFirstName() {
        return firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public Integer getAge(){ return age;}

    public void setAge(Integer newValue){
        this.age = newValue;
    }

    public String getHeight() {
        return height;
    }

    public void setHeight(String height) {
        this.height = height;
    }

    public Double getWeight() {
        return weight;
    }

    public void setWeight(Double weight) {
        this.weight = weight;
    }

    @Override
    public String toString() {
        return new StringBuilder()
                .append("\nMy first name is " + firstName)
                .append("\nMy last name is " + lastName)
                .toString();
    }
}
